changes from reviews
