+++ cmd/palomad/conifg.go
@@ -1,0 +1,13 @@
+package main
+
+import tmcfg "github.com/cometbft/cometbft/config"
+
+func initCometBFTConfig() *tmcfg.Config {
+	cfg := tmcfg.DefaultConfig()
+
+	// these values put a higher strain on node memory
+	// cfg.P2P.MaxNumInboundPeers = 100
+	// cfg.P2P.MaxNumOutboundPeers = 40
+
+	return cfg
+}
